+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Update `LinkBox` component.
+- Feature: Create `VerticalCard` component.
 
 ## [23.8.0] - 2018-10-26